# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateSalePoint:
    properties:
      branch_id:
        type: string
      name:
        type: string
    type: object
...
  models.SalePoint:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateSalePoint:
    properties:
      branch_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/sale_point:
    get:
      consumes:
      - application/json
      description: Get a list of sale points with optional filtering.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of sale points
          schema:
            items:
              $ref: '#/definitions/models.SalePoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a list of sale points
      tags:
      - sale_point
    post:
      consumes:
      - application/json
      description: Create a new sale point in the market system.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Sale point information
        in: body
        name: salePoint
        required: true
        schema:
          $ref: '#/definitions/models.CreateSalePoint'
      produces:
      - application/json
      responses:
        "201":
          description: Created sale point
          schema:
            $ref: '#/definitions/models.SalePoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new sale point
      tags:
      - sale_point
  /v1/sale_point/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing sale point.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Sale Point ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a sale point
      tags:
      - sale_point
    get:
      consumes:
      - application/json
      description: Get sale point details by its ID.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Sale Point ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sale point details
          schema:
            $ref: '#/definitions/models.SalePoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Sale point not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a sale point by ID
      tags:
      - sale_point
    put:
      consumes:
      - application/json
      description: Update an existing sale point.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Sale Point ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated sale point information
        in: body
        name: salePoint
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSalePoint'
      produces:
      - application/json
      responses:
        "202":
          description: Updated sale point
          schema:
            $ref: '#/definitions/models.SalePoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Sale point not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a sale point
      tags:
      - sale_point
...
swagger: "2.0"